+++ src/main/java/com/example/coditas/product/entity/ProductRequestMapping.java
@@ -10,7 +10,7 @@ public class ProductRequestMapping {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne @JoinColumn(name = "request_id", nullable = false)
     private ProductRequest request;
@@ -19,5 +19,5 @@ public class ProductRequestMapping {
     private Product product;
 
     @Column(nullable = false)
-    private Integer quantity = 0;
+    private Long quantity = 0L;
 }
